// apps/web/src/lib/api/model/judgment-result-verdict.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * edcurate-api
 * OpenAPI spec version: 0.1.0
 */

export type JudgmentResultVerdict = typeof JudgmentResultVerdict[keyof typeof JudgmentResultVerdict];


export const JudgmentResultVerdict = {
  use_it: 'use_it',
  adapt_it: 'adapt_it',
  skip_it: 'skip_it',
} as const;