// apps/web/src/lib/api/model/metric-result-rating.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * edcurate-api
 * OpenAPI spec version: 0.1.0
 */

export type MetricResultRating = typeof MetricResultRating[keyof typeof MetricResultRating];


export const MetricResultRating = {
  strong: 'strong',
  adequate: 'adequate',
  weak: 'weak',
} as const;